# .agents/resources/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ClearRequest": {
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          },
          "nodes": {
            "items": {
              "type": "string"
            },
            "title": "Nodes",
            "type": "array"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "required": [
          "graph",
          "nodes"
        ],
        "title": "ClearRequest",
        "type": "object"
      },
      "ColumnRefEdge": {
        "description": "Edge that maps a named output column to a named input.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source_node": {
            "title": "Source Node",
            "type": "string"
          },
          "source_output": {
            "title": "Source Output",
            "type": "string"
          },
          "target_input": {
            "title": "Target Input",
            "type": "string"
          },
          "target_node": {
            "title": "Target Node",
            "type": "string"
          },
          "type": {
            "const": "column_ref",
            "default": "column_ref",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_node",
          "target_node",
          "source_output",
          "target_input"
        ],
        "title": "ColumnRefEdge",
        "type": "object"
      },
...
      "ExecutionRequest": {
        "description": "Request to execute a graph (or a subset of its nodes).",
        "properties": {
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          },
          "graph": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph"
          },
          "nodes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nodes"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "title": "ExecutionRequest",
        "type": "object"
      },
...
      "GraphState": {
        "description": "Complete graph state with nodes and edges.",
        "properties": {
          "edges": {
            "items": {
              "discriminator": {
                "mapping": {
                  "column_ref": "#/components/schemas/ColumnRefEdge",
                  "positional": "#/components/schemas/PositionalEdge"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ColumnRefEdge"
                },
                {
                  "$ref": "#/components/schemas/PositionalEdge"
                }
              ]
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeState"
            },
            "title": "Nodes",
            "type": "array"
          },
          "published_inputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedInput"
            },
            "title": "Published Inputs",
            "type": "array"
          },
          "published_outputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedOutput"
            },
            "title": "Published Outputs",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "GraphState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeState": {
        "description": "A single node in the processing graph.",
        "properties": {
          "collapsed": {
            "default": false,
            "title": "Collapsed",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_templates": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Output Templates",
            "type": "object"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "position": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Position",
            "type": "array"
          },
          "published_inputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedInput"
            },
            "title": "Published Inputs",
            "type": "array"
          },
          "published_outputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedOutput"
            },
            "title": "Published Outputs",
            "type": "array"
          },
          "resources": {
            "additionalProperties": true,
            "title": "Resources",
            "type": "object"
          },
          "sub_workflow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphState"
              },
              {
                "type": "null"
              }
            ]
          },
          "sub_workflow_readonly_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Workflow Readonly Reason"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "tool_name",
          "position",
          "parameters"
        ],
        "title": "NodeState",
        "type": "object"
      },
...
      "PositionalEdge": {
        "description": "Edge that connects nodes by positional index.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "positional_index": {
            "title": "Positional Index",
            "type": "integer"
          },
          "source_node": {
            "title": "Source Node",
            "type": "string"
          },
          "target_node": {
            "title": "Target Node",
            "type": "string"
          },
          "type": {
            "const": "positional",
            "default": "positional",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_node",
          "target_node",
          "positional_index"
        ],
        "title": "PositionalEdge",
        "type": "object"
      },
      "PublishedInput": {
        "description": "An internal sub-workflow field exposed as an outer input pin.",
        "properties": {
          "default": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "internal_field": {
            "title": "Internal Field",
            "type": "string"
          },
          "internal_node_id": {
            "title": "Internal Node Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "parameter",
              "input"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema"
          }
        },
        "required": [
          "name",
          "internal_node_id",
          "internal_field",
          "kind"
        ],
        "title": "PublishedInput",
        "type": "object"
      },
      "PublishedOutput": {
        "description": "An internal sub-workflow output exposed as an outer output pin.",
        "properties": {
          "internal_node_id": {
            "title": "Internal Node Id",
            "type": "string"
          },
          "internal_output": {
            "title": "Internal Output",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema"
          }
        },
        "required": [
          "name",
          "internal_node_id",
          "internal_output"
        ],
        "title": "PublishedOutput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BioImageFlow Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/execution/clear": {
      "post": {
        "operationId": "clear_execution_api_v1_execution_clear_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clear Execution Api V1 Execution Clear Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Execution",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/v1/execution/run": {
      "post": {
        "operationId": "run_execution_api_v1_execution_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Execution",
        "tags": [
          "execution"
        ]
      }
    },
...
  }
}